fix: unionBy method when merge config
