+++ examples/with-rax/src/app.tsx
@@ -17,5 +17,12 @@ runApp({
       console.log('app hide...');
     },
 
+    // 获取初始数据
+    getInitialData: async () => {
+      return {
+        a: 1,
+        b: 2
+      };
+    }
   }
 });
