+++ packages/build-app-helpers/src/index.ts
@@ -3,4 +3,5 @@ export { default as getRoutesByAppJson } from './getRoutesByAppJson';
 export { default as formatPath } from './formatPath';
 export { default as validation } from './validation';
 export { default as unionBy } from './unionBy';
+export { default as injectTransformRuntime } from './injectTransformRuntime';
