+++ packages/plugin-react-app/src/userConfig/injectBabel.js
@@ -1,4 +1,4 @@
-const { formatPath } = require('build-app-helpers');
+const { injectTransformRuntime } = require('build-app-helpers');
 const chalk = require('chalk');
 const addBablePlugins = require('../utils/addBabelPlugins');
 
@@ -9,31 +9,7 @@ module.exports = (config, injectBabel, context) => {
   }
   console.log(chalk.cyan('Detected  you are using injectBabel, please use polyfill, Visit https://ice.work/docs/guide/basic/build.'));
   if (injectBabel === 'runtime') {
-    ['jsx', 'tsx'].forEach((rule) => {
-      config.module
-        .rule(rule)
-        .use('babel-loader')
-        .tap((options) => {
-          // get @babel/plugin-transform-runtime
-          const babelPlugins = options.plugins || [];
-          const targetPlugin = formatPath('@babel/plugin-transform-runtime');
-          const plguinOption = {
-            corejs: false,
-            helpers: true,
-            regenerator: true,
-            useESModules: false,
-          };
-          const plugins = babelPlugins.map((plugin) => {
-            if (typeof plugin === 'string' && formatPath(plugin).indexOf(targetPlugin) > -1
-              || Array.isArray(plugin) && formatPath(plugin[0]).indexOf(targetPlugin) > -1 ) {
-              return [Array.isArray(plugin) ? plugin[0] : plugin, plguinOption];
-            } else {
-              return [require.resolve('@babel/plugin-transform-runtime'), plguinOption];
-            }
-          });
-          return Object.assign(options, { plugins });
-        });
-    });
+    injectTransformRuntime(config);
   } else if (injectBabel === 'polyfill') {
     const entries = config.toConfig().entry;
     const rule = config.module.rule('polyfill').test(/\.jsx?|\.tsx?$/);
